requirements.txt
Django==1.6.5
South==1.0
googleapiclient==1.2
httplib2==0.9
oauth2client==1.2
oauthlib==0.6.3
psycopg2==2.5.3
pysqlite==2.6.3
python-dateutil==2.2
python-openid==2.2.5
python-social-auth==0.1.26
pytz==2014.4
pyzipcode==0.4
requests==2.3.0
requests-oauthlib==0.4.1
simplejson==3.6.2
six==1.7.3
tzlocal==1.1.1
uritemplate==0.6
urllib3==1.9
wsgiref==0.1.2
